Splitting up the ITs Adding IntegrationTestRule to manage DockerComposeRule, and Morphia datastore
