+++ minianalytics-api/src/test/java/amcmurray/bw/integrationTests/MentionIT.java
@@ -14,39 +14,20 @@ import java.util.Date;
 
 import org.junit.After;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.ClassRule;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.mongodb.morphia.Datastore;
-import org.mongodb.morphia.Morphia;
 
-import com.mongodb.MongoClient;
-import com.palantir.docker.compose.DockerComposeRule;
-import com.palantir.docker.compose.configuration.ProjectName;
-import com.palantir.docker.compose.configuration.ShutdownStrategy;
-import com.palantir.docker.compose.connection.waiting.HealthChecks;
-
-import amcmurray.bw.QueryRequestDTO;
 import amcmurray.bw.exceptions.QueryExceptions;
 import amcmurray.bw.twitterdomainobjects.Mention;
 import amcmurray.bw.twitterdomainobjects.MentionType;
 import amcmurray.bw.twitterdomainobjects.Query;
-import io.restassured.http.ContentType;
 
 @Category(IntegrationTest.class)
-public class miniananalyticsApiIT {
+public class MentionIT {
 
-    private static final String MONGO_CONTAINER_NAME = "mongodbtest";
-    private static final int MONGO_INTERNAL_PORT = 27017;
-    private static int MONGO_EXTERNAL_PORT;
-
-    private static final String API_CONTAINER_NAME = "minianalyticsapitest";
-    private static final int API_INTERNAL_PORT = 8081;
-    private static int API_EXTERNAL_PORT;
-
-    private static final String DATABASE_NAME = "minianalytics";
-    private static Datastore datastore;
+    @ClassRule
+    public static IntegrationTestRule rule = IntegrationTestRule.INSTANCE;
 
     private final Date testDate1 = Timestamp.valueOf(LocalDateTime.parse("2018-08-31T18:30:00.000",
             DateTimeFormatter.ISO_DATE_TIME));
@@ -66,36 +47,6 @@ public class miniananalyticsApiIT {
     private final Mention mention3 = new Mention("789hij", 1, MentionType.TWITTER,
             "testAuthor1", "eine test Suche", testDate2, "de", 0);
 
-    @ClassRule
-    public static DockerComposeRule docker = DockerComposeRule.builder()
-            .file("src/test/resources/docker-compose-tests.yml")
-            .projectName(ProjectName.random())
-            .waitingForService(API_CONTAINER_NAME, HealthChecks.toHaveAllPortsOpen())
-            .saveLogsTo("target/docker-compose-test-logs")
-            .shutdownStrategy(ShutdownStrategy.GRACEFUL)
-            .build();
-
-    @BeforeClass
-    public static void initialize() {
-        //getting external port for API service
-        API_EXTERNAL_PORT = docker.containers()
-                .container(API_CONTAINER_NAME)
-                .port(API_INTERNAL_PORT).getExternalPort();
-
-        MONGO_EXTERNAL_PORT = docker.containers()
-                .container(MONGO_CONTAINER_NAME)
-                .port(MONGO_INTERNAL_PORT).getExternalPort();
-
-        //using morphia to create new datastore
-        final Morphia morphia = new Morphia();
-        morphia.mapPackage("amcmurray.bw.twitterdomainobjects");
-
-        //datastore uses mongo instance
-        datastore = morphia.createDatastore(
-                new MongoClient("localhost", MONGO_EXTERNAL_PORT), DATABASE_NAME);
-        datastore.ensureIndexes();
-    }
-
     @Before
     public void setup() {
         addQueriesToDB();
@@ -105,69 +56,6 @@ public class miniananalyticsApiIT {
     @After
     public void tearDown() {
         dropDB();
-    }
-
-    @Test
-    public void viewQueryById_getsExpectedQuery() {
-        with().get(createURLWithPort("/queries/0"))
-                .then().assertThat()
-                .statusCode(200)
-                .body("id", equalTo(query1.getId()))
-                .body("text", equalTo(query1.getText()));
-    }
-
-    @Test
-    public void viewNonexistentQueryById_throwsQueryNotFoundException() {
-        with().get(createURLWithPort("/queries/-1"))
-
-                .then().assertThat()
-                .statusCode(404)
-                .body("exception",
-                        equalTo(QueryExceptions.QueryNotFoundException.class.getName()));
-    }
-
-    @Test
-    public void viewMultipleQueries_returnsMultiple() {
-        with().get(createURLWithPort("/queries"))
-
-                .then().assertThat()
-                .statusCode(200)
-                .body("id", hasSize(2))
-                .body("id[0]", equalTo(query1.getId()))
-                .body("text[0]", equalTo(query1.getText()))
-                .body("id[1]", equalTo(query2.getId()))
-                .body("text[1]", equalTo(query2.getText()));
-    }
-
-    @Test
-    public void addNewQuery_returnsExpectedQuery() {
-        QueryRequestDTO queryRequestDTO = new QueryRequestDTO("new search", "pl");
-
-        with().contentType(ContentType.JSON)
-                .body(queryRequestDTO)
-                .when()
-                .post((createURLWithPort("/queries")))
-
-                .then().assertThat()
-                .statusCode(200)
-                .body("id", equalTo(query2.getId() + 1))
-                .body("text", equalTo(queryRequestDTO.getSearch()))
-                .body("language", equalTo(queryRequestDTO.getLanguage()));
-    }
-
-    @Test
-    public void addNewQueryWithBlankSearch_throwsQuerySearchNullException() {
-        QueryRequestDTO queryRequestDTO = new QueryRequestDTO("", "en");
-
-        with().contentType(ContentType.JSON)
-                .body(queryRequestDTO)
-                .when()
-                .post((createURLWithPort("/queries")))
-
-                .then().assertThat()
-                .statusCode(400)
-                .body("exception",
-                        equalTo(QueryExceptions.QuerySearchNullException.class.getName()));
     }
 
     @Test
@@ -309,22 +197,22 @@ public class miniananalyticsApiIT {
     }
 
     private String createURLWithPort(String uri) {
-        return "http://localhost:" + API_EXTERNAL_PORT + uri;
+        return "http://localhost:" + rule.getApiExternalPort() + uri;
     }
 
     private void addQueriesToDB() {
-        datastore.save(query1);
-        datastore.save(query2);
+        rule.getDatastore().save(query1);
+        rule.getDatastore().save(query2);
     }
 
     private void dropDB() {
-        datastore.getDB().getCollection("savedQueries").drop();
-        datastore.getDB().getCollection("savedMentions").drop();
+        rule.getDatastore().getDB().getCollection("savedQueries").drop();
+        rule.getDatastore().getDB().getCollection("savedMentions").drop();
     }
 
     private void addMentionsToDB() {
-        datastore.save(mention1);
-        datastore.save(mention2);
-        datastore.save(mention3);
+        rule.getDatastore().save(mention1);
+        rule.getDatastore().save(mention2);
+        rule.getDatastore().save(mention3);
     }
 }
